Add files via upload
The work here is to predict customer eligibility for the loan process of a bank.  To solve this problem we uses Random forest regression algorithm to build the model for prediction. the training data set have missing data and categorical data, to be able to use this data in our machine learning algorithm need to do the preprocessing step  to deal with the missing value and to transform the categorical data to numeric data for our model.